@@ -19,56 +19,56 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testWillParseFlagsOutOfTheClass()
     {
-        $fixture = new Fixture(self::USER.' (local)', 'user', array(), null);
+        $fixture = new Fixture(self::USER.' (local)', 'user', [], null);
 
         $this->assertEquals(self::USER, $fixture->getClass());
     }
 
     public function testWillParseFlagsOutOfTheName()
     {
-        $fixture = new Fixture(self::USER, 'user (local)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (local)', [], null);
 
         $this->assertEquals('user', $fixture->getName());
     }
 
     public function testIsLocalWithLocalClassFlag()
     {
-        $fixture = new Fixture(self::USER.' (local)', 'user', array(), null);
+        $fixture = new Fixture(self::USER.' (local)', 'user', [], null);
 
         $this->assertTrue($fixture->isLocal());
     }
 
     public function testIsLocalWithLocalNameFlag()
     {
-        $fixture = new Fixture(self::USER, 'user (local)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (local)', [], null);
 
         $this->assertTrue($fixture->isLocal());
     }
 
     public function testIsNotLocalWithNeitherClassNorNameFlag()
     {
-        $fixture = new Fixture(self::USER, 'user', array(), null);
+        $fixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertFalse($fixture->isLocal());
     }
 
     public function testIsTemplateWithTemplateNameFlag()
     {
-        $fixture = new Fixture(self::USER, 'user (template)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (template)', [], null);
 
         $this->assertTrue($fixture->isTemplate());
     }
 
     public function testIsNotTemplateWithoutTemplateNameFlag()
     {
-        $fixture = new Fixture(self::USER, 'user', array(), null);
+        $fixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertFalse($fixture->isTemplate());
     }
 
     public function testIsNotTemplateWithExtendsNameFlag($value = '')
     {
-        $fixture = new Fixture(self::USER, 'user (extends user_template)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (extends user_template)', [], null);
 
         $this->assertFalse($fixture->isTemplate());
     }
@@ -79,15 +79,15 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
      */
     public function testExtendTemplateRequiresThatTheArgumentIsATemplate()
     {
-        $fixture1 = new Fixture(self::USER, 'user1', array(), null);
-        $fixture2 = new Fixture(self::USER, 'user2', array(), null);
+        $fixture1 = new Fixture(self::USER, 'user1', [], null);
+        $fixture2 = new Fixture(self::USER, 'user2', [], null);
         $fixture1->extendTemplate($fixture2);
     }
 
     public function testExtendTemplateWillMapUnsetPropertiesOnTheFixture()
     {
-        $template = new Fixture(self::USER, 'user_full (template)', array('name' => 'John Doe', 'email' => 'john@doe.org'), null);
-        $fixture = new Fixture(self::USER, 'user', array('email' => 'jane@doe.org'), null);
+        $template = new Fixture(self::USER, 'user_full (template)', ['name' => 'John Doe', 'email' => 'john@doe.org'], null);
+        $fixture = new Fixture(self::USER, 'user', ['email' => 'jane@doe.org'], null);
 
         $fixture->extendTemplate($template);
         $properties = $fixture->getProperties();
@@ -96,8 +96,8 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testExtendTemplateWillNotMapSetPropertiesOnTheFixture()
     {
-        $template = new Fixture(self::USER, 'user_full (template)', array('name' => 'John Doe', 'email' => 'john@doe.org'), null);
-        $fixture = new Fixture(self::USER, 'user', array('email' => 'jane@doe.org'), null);
+        $template = new Fixture(self::USER, 'user_full (template)', ['name' => 'John Doe', 'email' => 'john@doe.org'], null);
+        $fixture = new Fixture(self::USER, 'user', ['email' => 'jane@doe.org'], null);
 
         $fixture->extendTemplate($template);
         $properties = $fixture->getProperties();
@@ -106,36 +106,36 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testGetExtensionsReturnsAListOfAllTemplateNamesTheFixtureExtends()
     {
-        $fixture = new Fixture(self::USER, 'user (extends user_name, extends user_email)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (extends user_name, extends user_email)', [], null);
 
-        $this->assertEquals(array('user_name', 'user_email'), $fixture->getExtensions());
+        $this->assertEquals(['user_name', 'user_email'], $fixture->getExtensions());
     }
 
     public function testHasExtensionsIsFalseWhenNoExtensionsExist()
     {
-        $fixture = new Fixture(self::USER, 'user', array(), null);
+        $fixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertFalse($fixture->hasExtensions());
     }
 
     public function testHasExtensionsIsTrueWhenExtensionsExist()
     {
-        $fixture = new Fixture(self::USER, 'user (extends user_name, extends user_email)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (extends user_name, extends user_email)', [], null);
 
         $this->assertTrue($fixture->hasExtensions());
     }
 
     public function testGetPropertiesWillReturnOnlyBasicValueProperties()
     {
-        $fixture = new Fixture(self::USER, 'user', array('name' => 'John Doe', 'email' => 'john@doe.org', '__construct' => array('1', '2'), '__set' => 'setterFunc'), null);
+        $fixture = new Fixture(self::USER, 'user', ['name' => 'John Doe', 'email' => 'john@doe.org', '__construct' => ['1', '2'], '__set' => 'setterFunc'], null);
 
         $properties = $fixture->getProperties();
-        $this->assertEquals(array('name' => $properties['name'], 'email' => $properties['email']), $fixture->getProperties());
+        $this->assertEquals(['name' => $properties['name'], 'email' => $properties['email']], $fixture->getProperties());
     }
 
     public function testHasClassFlagWillReturnIfClassFLagExists()
     {
-        $fixture = new Fixture(self::USER.' (local)', 'user', array(), null);
+        $fixture = new Fixture(self::USER.' (local)', 'user', [], null);
 
         $this->assertTrue($fixture->hasClassFlag('local'));
         $this->assertFalse($fixture->hasClassFlag('badname'));
@@ -143,7 +143,7 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testHasNameFlagWillReturnIfNameFLagExists()
     {
-        $fixture = new Fixture(self::USER, 'user (local)', array(), null);
+        $fixture = new Fixture(self::USER, 'user (local)', [], null);
 
         $this->assertTrue($fixture->hasNameFlag('local'));
         $this->assertFalse($fixture->hasNameFlag('badname'));
@@ -151,43 +151,43 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testGetConstructorMethodWillReturnTheMethodName()
     {
-        $fixture = new Fixture(self::USER, 'user', array('__construct' => array('create' => array('1', '2', '3'))), null);
+        $fixture = new Fixture(self::USER, 'user', ['__construct' => ['create' => ['1', '2', '3']]], null);
 
         $this->assertEquals('create', $fixture->getConstructorMethod());
     }
 
     public function testGetConstructorArgsWillReturnTheArgumentsList()
     {
-        $fixture = new Fixture(self::USER, 'user', array('__construct' => array('create' => array('1', '2', '3'))), null);
+        $fixture = new Fixture(self::USER, 'user', ['__construct' => ['create' => ['1', '2', '3']]], null);
 
-        $this->assertEquals(array('1', '2', '3'), $fixture->getConstructorArgs());
+        $this->assertEquals(['1', '2', '3'], $fixture->getConstructorArgs());
     }
 
     public function testShouldUseConstructorWillReturnTrueIfThereIsNoConstructorInTheSpec()
     {
-        $fixture = new Fixture(self::USER, 'user', array(), null);
+        $fixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertTrue($fixture->shouldUseConstructor());
     }
 
     public function testShouldUseConstructorWillReturnFalseIfTheConstructorSpecIsFalse()
     {
-        $fixture = new Fixture(self::USER, 'user', array('__construct' => false), null);
+        $fixture = new Fixture(self::USER, 'user', ['__construct' => false], null);
 
         $this->assertFalse($fixture->shouldUseConstructor());
     }
 
     public function testShouldUseConstructorWillReturnTrueIfTheConstructorSpecIsDefined()
     {
-        $fixture = new Fixture(self::USER, 'user', array('__construct' => array('1', '2')), null);
+        $fixture = new Fixture(self::USER, 'user', ['__construct' => ['1', '2']], null);
 
         $this->assertTrue($fixture->shouldUseConstructor());
     }
 
     public function testHasCustomerSetterWillReturnIfTheSpecDefinesACustomSetter()
     {
-        $setFixture = new Fixture(self::USER, 'user', array('__set' => 'setterFunc'), null);
-        $noSetFixture = new Fixture(self::USER, 'user', array(), null);
+        $setFixture = new Fixture(self::USER, 'user', ['__set' => 'setterFunc'], null);
+        $noSetFixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertTrue($setFixture->hasCustomSetter());
         $this->assertFalse($noSetFixture->hasCustomSetter());
@@ -195,8 +195,8 @@ class FixtureTest extends \PHPUnit_Framework_TestCase
 
     public function testGetCustomSetterWillReturnTheCustomSetterValue()
     {
-        $setFixture = new Fixture(self::USER, 'user', array('__set' => 'setterFunc'), null);
-        $noSetFixture = new Fixture(self::USER, 'user', array(), null);
+        $setFixture = new Fixture(self::USER, 'user', ['__set' => 'setterFunc'], null);
+        $noSetFixture = new Fixture(self::USER, 'user', [], null);
 
         $this->assertEquals('setterFunc', $setFixture->getCustomSetter());
         $this->assertNull($noSetFixture->getCustomSetter());
