@@ -1,14 +1,14 @@
 <?php
 
-return array(
-    'Nelmio\Alice\support\models\User' => array(
-        'user1' => array(
+return [
+    'Nelmio\Alice\support\models\User' => [
+        'user1' => [
             'username' => '<identity($fake("upperCaseProvider", null, "John Doe"))>',
             'fullname' => '<upperCaseProvider("John Doe")>',
-        ),
-        'user2' => array(
+        ],
+        'user2' => [
             'username' => $fake('identity', null, $fake('upperCaseProvider', null, 'John Doe')),
             'fullname' => $fake('upperCaseProvider', null, 'John Doe'),
-        ),
-    ),
-);
+        ],
+    ],
+];
